debuging and testing api

+++ src/controllers/playlist.controller.js
@@ -5,10 +5,13 @@ import { ApiError } from '../utills/apiError.js';
 
 const createPlaylist = asyncHandler(async (req, res) => {
   const { name, description } = req.body;
+  console.log(name, description);
   if (!name) {
     throw new ApiError(400, 'playlist name is missing');
   }
+  console.log(req.user?._id);
   try {
+    console.log(name, description, req.user?._id);
     const playlist = await Playlist.create({
       name,
       description,
@@ -16,9 +19,9 @@ const createPlaylist = asyncHandler(async (req, res) => {
     });
     return res
       .status(200)
-      .json(new ApiResponse(200, playlist, 'playlist created successfully'));
+      .json(new ApiResponse(200, {}, 'playlist created successfully'));
   } catch (error) {
-    throw new ApiError(400, 'error while creating playlist :', error.message);
+    throw new ApiError(400, error.message);
   }
 });
 
@@ -30,6 +33,9 @@ const getUserPlaylists = asyncHandler(async (req, res) => {
   }
   try {
     const userPlaylist = await Playlist.find({ owner: userId });
+    if (!userPlaylist.length) {
+      throw new ApiError(400, 'no any playlist is created here');
+    }
     return res
       .status(200)
       .json(
@@ -47,6 +53,9 @@ const getPlaylistById = asyncHandler(async (req, res) => {
   }
   try {
     const playlist = await Playlist.findById(playlistId).populate('videos');
+    if (!playlist) {
+      throw new ApiError(404, 'playlist not found');
+    }
     return res
       .status(200)
       .json(new ApiResponse(200, playlist, 'playlist fetched successfully'));
@@ -58,18 +67,21 @@ const getPlaylistById = asyncHandler(async (req, res) => {
 
 const addVideoToPlaylist = asyncHandler(async (req, res) => {
   const { playlistId, videoId } = req.params;
+
   if (!playlistId || !videoId) {
     throw new ApiError(400, 'playlist id or videoid is missing');
   }
   try {
-    const addVideoPlaylist = await Playlist.findByIdAndUpdate(
-      playlistId,
-      { $push: { videos: videoId } },
-      { new: true }
-    );
+    const playlist = await Playlist.findById(playlistId);
+    if (playlist?.videos.includes(String(videoId))) {
+      throw new ApiError(400, 'video is alredy added in playlist');
+    }
+    playlist?.videos.push(videoId);
+
+    await playlist.save();
     return res
       .status(200)
-      .json(new ApiResponse(200, addVideoPlaylist, 'video added successfully'));
+      .json(new ApiResponse(200, {}, 'video added successfully'));
   } catch (error) {
     throw new ApiError(400, error.message);
   }
@@ -88,9 +100,7 @@ const removeVideoFromPlaylist = asyncHandler(async (req, res) => {
     );
     return res
       .status(200)
-      .json(
-        new ApiResponse(200, addVideoPlaylist, 'video removed successfully')
-      );
+      .json(new ApiResponse(200, {}, 'video removed successfully'));
   } catch (error) {
     throw new ApiError(400, error.message);
   }
@@ -102,7 +112,12 @@ const deletePlaylist = asyncHandler(async (req, res) => {
   if (!playlistId) {
     throw new ApiError(400, 'playlist id is missing');
   }
+
   try {
+    const playlist = await Playlist.findById(playlistId);
+    if (String(playlist?.owner) !== String(req.user?._id)) {
+      throw new ApiError(404, 'unothorized request');
+    }
     await Playlist.findByIdAndDelete(playlistId);
     return res
       .status(200)
